+++ src/main/java/com/example/lucky7/config/SecurityConfig.java
@@ -1,5 +1,6 @@
 package com.example.lucky7.config;
 
+import com.example.lucky7.config.jwt.JwtAuthenticationFilter;
 import com.example.lucky7.domain.user.enums.UserRole;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
@@ -44,6 +45,8 @@ public class SecurityConfig {
                         .requestMatchers("/health").permitAll()
                         .requestMatchers(request -> request.getRequestURI().startsWith("/auth")).permitAll()
                         .requestMatchers(request -> request.getRequestURI().startsWith("/admin")).hasAuthority(UserRole.Authority.ADMIN)
+                        .requestMatchers(request -> request.getRequestURI().startsWith("/swagger-ui")).permitAll()
+                        .requestMatchers(request -> request.getRequestURI().startsWith("/v3/api-docs")).permitAll()
                         .anyRequest().authenticated()
                 )
                 .build();
