+++ src/main/java/com/example/lucky7/config/jwt/JwtAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.example.lucky7.config;
+package com.example.lucky7.config.jwt;
 
 import com.example.lucky7.domain.common.dto.AuthUser;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
